Lab3/pzpi-22-10-zabolotnyi-daniil-lab3/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0
PyJWT>=2.6.0
psycopg2-binary>=2.9.0
firebase-admin>=6.0.0
requests>=2.28.0
openpyxl>=3.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0 